move client queries and mutations to client folder
